Create bin and update readme
