+++ apps/ble-reader/src/bin.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env/ node
+
+import { program } from "./main.js";
+
+program.parse();
